// packages/tbd-store/clients/blh/bet-live-hypotheticals-response-types.ts
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
};

/**  The Resulted Bet */
export type BetResult = {
  __typename?: "BetResult";
  legs: Array<Leg>;
  result: Result;
  /**  The result of this bet */
  resultType: ResultType;
};

export enum BetType {
  CANADIAN = "CANADIAN",
  DOUBLE = "DOUBLE",
  EIGHTEENFOLD = "EIGHTEENFOLD",
  EIGHTFOLD = "EIGHTFOLD",
  ELEVENFOLD = "ELEVENFOLD",
  FIFTEENFOLD = "FIFTEENFOLD",
  FIVEFOLD = "FIVEFOLD",
  FOURFOLD = "FOURFOLD",
  FOURTEENFOLD = "FOURTEENFOLD",
  GOLIATH = "GOLIATH",
  HEINZ = "HEINZ",
  LUCKY_15 = "LUCKY_15",
  LUCKY_31 = "LUCKY_31",
  LUCKY_63 = "LUCKY_63",
  MANUAL = "MANUAL",
  NINEFOLD = "NINEFOLD",
  NINETEENFOLD = "NINETEENFOLD",
  PATENT = "PATENT",
  SEVENFOLD = "SEVENFOLD",
  SEVENTEENFOLD = "SEVENTEENFOLD",
  SINGLE = "SINGLE",
  SIXFOLD = "SIXFOLD",
  SIXTEENFOLD = "SIXTEENFOLD",
  SUPER_HEINZ = "SUPER_HEINZ",
  TENFOLD = "TENFOLD",
  THIRTEENFOLD = "THIRTEENFOLD",
  TREBLE = "TREBLE",
  TRIXIE = "TRIXIE",
  TWELVEFOLD = "TWELVEFOLD",
  TWENTYFIVEFOLD = "TWENTYFIVEFOLD",
  TWENTYFOLD = "TWENTYFOLD",
  TWENTYFOURFOLD = "TWENTYFOURFOLD",
  TWENTYONEFOLD = "TWENTYONEFOLD",
  TWENTYTHREEFOLD = "TWENTYTHREEFOLD",
  TWENTYTWOFOLD = "TWENTYTWOFOLD",
  YANKEE = "YANKEE",
}
// ...
export type InputBet = {
  betType: BetType;
  legs: Array<InputLeg>;
};

/**  Input Bet Leg used for Result Calculation */
export type InputLeg = {
  legNumber: Scalars["ID"]["input"];
  runners: Array<InputRunner>;
};

/**  Input Leg Runner used for Result Calculation */
export type InputRunner = {
  id: Scalars["ID"]["input"];
  /**  SPORTEX Id of the Runner */
  marketId: Scalars["ID"]["input"];
  /**  Sportsbook Id of the Market */
  result: InputMaybe<Result>;
};
// ...
export enum Result {
  /**  Winning Bet/Leg */
  LOSE = "LOSE",
  /**  Voided Bet/Leg */
  PENDING = "PENDING",
  /**  Losing Bet/Leg */
  VOID = "VOID",
  WIN = "WIN",
}

/**  How the Bet/Leg Result was Calculated */
export enum ResultType {
  CONFIRMED = "CONFIRMED",
  /**  Result received from either Catalog or Channels */
  POTENTIAL = "POTENTIAL",
  /**  Potential Result based on Real World Data */
  UNKNOWN = "UNKNOWN",
}